Hide public Solana Summit agenda
