+++ src/app/solana-summit-germany/page.tsx
@@ -290,13 +290,9 @@ function AgendaIntro() {
             </div>
           ))}
         </div>
-        <SummitCta
-          href="/solana-summit-germany/agenda"
-          variant="solid"
-          className="mt-12 min-h-[56px] px-10 text-[14px]"
-        >
-          View Full Agenda
-        </SummitCta>
+        <p className="mx-auto mt-12 max-w-[460px] text-sm font-semibold uppercase tracking-[0.16em] text-white/55">
+          Full agenda will be announced one week before the summit.
+        </p>
       </div>
     </SummitReveal>
   );
